see #8902 - fix compilation warnings
git-svn-id: http://josm.openstreetmap.de/svn/trunk@6142 0c6e7542-c601-0410-84e7-c038aed88b3b

+++ src/org/openstreetmap/josm/tools/Diff.java
@@ -815,7 +815,7 @@ public class Diff {
             realindexes = new int[buffered_lines];
 
             for (int i = 0; i < data.length; ++i) {
-                Integer ir = (Integer)h.get(data[i]);
+                Integer ir = h.get(data[i]);
                 if (ir == null) {
                     h.put(data[i],new Integer(equivs[i] = equiv_max++));
                 } else {
